refactor(User): 유저 4-layer적용
